clean defaults
